// frontend/src/app/core/api/generated/model/product-request.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductRequest { 
    sku: string;
    name: string;
    description?: string;
    price: number;
    active: boolean;
    /**
     * Initial minimum for MAIN when creating a product; catalog updates do not modify warehouse settings
     */
    minimumStock?: number;
}